requirements.txt
amqp==5.2.0
asgiref==3.8.1
astroid==3.3.4
asttokens==2.4.1
billiard==4.2.1
black==24.8.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.1.1
diff-match-patch==20230430
dill==0.3.9
Django==5.1.1
django-import-export==4.1.1
executing==2.1.0
ipdb==0.13.13
ipython==8.28.0
isort==5.13.2
jedi==0.19.1
kombu==5.4.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pylint==3.3.1
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tablib==3.5.0
tomlkit==0.13.2
traitlets==5.14.3
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
